meta-mxt-rzv2l: app-usbcam-http: fix OpenCV downscaling of YUYV format
As it turns out cv::resize() doesn't work too well for YUYV formats.
The image color is messed up.

As such, we'll need to convert YUYV to RGB, downscale, then convert back
to YUYV.

Found this online:
   https://stackoverflow.com/questions/49964259/cpp-rgb-to-yuv422-conversion
Needed a bit of tweaking though.
It assigns UYVY, whereas we need YUYV.

Signed-off-by: Alexandru Ardelean <alex@shruggie.ro>