meta-mxt: kernel: replace Sony IMX296 with upstream version
The upstream version is way more stable than the one in the initial BSP.
It can be started and stopped multiple times at runtime.

The one from the initial BSP also uses the Simple ISP package.
The SimpleISP package doesn't work too well with all sensors yet (which is
why it's not being used).

Signed-off-by: Alexandru Ardelean <alex@shruggie.ro>