meta-mxt-rzv2l: add support Aptina/OnSemi AR0331/AR0330
Untested yet.

The driver is based on the structure of the imx219 driver.
The registers and controls were adapted from an older AR0330 driver found
in an nVIDIA tegra kernel
    https://nv-tegra.nvidia.com/r/gitweb?p=linux-3.10.git;a=commit;h=7a84cd9698c8c8ecb56b45a5a5ed20efeee6dca8

The AR0330 and AR0331 have the same register set.
This was confirmed by the driver found in TI's SDK for DM385 with RDK.
   https://www.ti.com/tool/TMDSCSK388
   https://www.ti.com/tool/IPNC-RDK-CSK  -- this SDK.

Signed-off-by: Alexandru Ardelean <aardelean@baylibre.com>