meta-mxt: add solution to make Sony IMX274 work on RZV2L
The Sony IMX274 sensor seems to be a bit finicky.
There are more registers that need tuning to get a stable stream of
1920x1080.

The current mode runs at 4K and crops a section of 1920x1080 in the middle.
The driver seems to need more tuning to actually get it to work with
more modes.

Signed-off-by: Alexandru Ardelean <alex@shruggie.ro>